requirements.txt
# Web Framework and Server
fastapi
uvicorn
pydantic

# HTTP client
httpx

# Database
sqlalchemy
psycopg2-binary==2.9.9

# Configuration
python-dotenv

# Reddit API
praw
requests

# Machine Learning - Inference
transformers
torch
torchvision
sentencepiece 
accelerate

# MLOps
mlflow
boto3

# Pydantic settings management
pydantic-settings

# Google GenAI
google-genai
